Add a notion of IMPORTANCE to differences.
IMPORTANCE = 1 is indicated by differing definitions, and that's of 'most' importance.
IMPORTANCE = 2 is indicated by an object being missing in one database, and that's a bit less critical.
IMPORTANCE = 3+ is indicated by an object being missing, and being a 'child' of an object with a lower
                IMPORTANCE value.  Thus, if a table is missing, then columns and such will get IMPORTANCE
                incremented because they're less crucial than the missing table.

It USUALLY suffices to check for ... WHERE importance in (1,2)